Refactor pydesigner package
