+++ pydesigner/fitting/dirs30.csv
