+++ pydesigner/plotting/motionplot.py
