+++ pydesigner/preprocessing/mrinfoutil.py
