+++ pydesigner/preprocessing/smoothing.py
