+++ pyproject.toml
@@ -2,6 +2,7 @@
 name = "pydesigner"
 version = "1.1.0"
 description = "A hands-free DTI, DKI, FBI and FBWM preprocessing pipeline"
+license = "Proprietary"
 authors = ["Siddhartha Dhiman <sid.kill3r@gmail.com>"]
 readme = "README.rst"
 
@@ -15,7 +16,10 @@ multiprocess = "^0.70.14"
 nibabel = "^5.1.0"
 dipy = "^1.7.0"
 cvxpy = "^1.3.1"
+matplotlib = "^3.7.1"
 
+[tool.poetry.scripts]
+pydesigner = "pydesigner.main:main"
 
 [build-system]
 requires = ["poetry-core"]
